+++ gatsby-config.js
@@ -87,6 +87,12 @@ module.exports = {
           },
         ],
       },
-    }
+    },{
+      resolve: `gatsby-plugin-google-analytics`,
+      options: {
+        trackingId: "UA-1322093-4",
+      },
+    },
+  ],
   ]
 }